Revert "Milestone 3: Events"
