+++ src/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import './App.css';
-import Calculator from './components/UI/Calculator';
+import Calculator from './components/Calculator';
 
 class App extends React.Component {
   constructor(props) {
