+++ src/components/Calculator.css
@@ -20,8 +20,3 @@
 .table .orange-colour {
   background: #f28f3b;
 }
-
-input {
-  background: none;
-  border: none;
-}
