--- src/components/logic/operate.js
@@ -1,27 +1,0 @@
-/* eslint import/no-extraneous-dependencies: ["error", {"peerDependencies": true}] */
-import Big from 'big.js';
-
-export default function operate(numberOne, numberTwo, operation) {
-  const one = Big(numberOne);
-  const two = Big(numberTwo);
-  if (operation === '+') {
-    return one.plus(two).toString();
-  }
-  if (operation === '-') {
-    return one.minus(two).toString();
-  }
-  if (operation === 'x') {
-    return one.times(two).toString();
-  }
-  if (operation === '÷') {
-    try {
-      return one.div(two).toString();
-    } catch (err) {
-      return "Can't divide by 0.";
-    }
-  }
-  if (operation === '%') {
-    return one.mod(two).toString();
-  }
-  throw Error(`Unknown operation '${operation}'`);
-}
